[rdkb][common][bsp][Fix overlayfs mount failure when dual-image is enabled and rootfs_data sharing is disabled]
[Description]
Fix overlayfs mount failure.

Root cause:
When dual-image is enabled in u-boot and rootfs_data sharing is
disabled, overlayfs must be mounted on /dev/fitrw instead. Previously,
the code always looked up the rootfs_data partition and mounted
overlayfs on it, regardless of this setting.

Fix:
Check whether "no-split-fitrw" exists to determine the correct mount
target:
- If it exists, rootfs_data sharing is enabled, so overlayfs mounts
  on the rootfs_data partition.
- If it does not exist, rootfs_data sharing is disabled, so overlayfs
  mounts on fitrw.

[Release-log]

+++ recipes-devtools/initrd/files/kernelv6-init_readonlyfs-emmc.sh
@@ -25,8 +25,9 @@ $MOUNT -n -t tmpfs tmpfs -o rw,nosuid,nodev,noexec,noatime /rdklogs
 [ -z "$CONSOLE" ] && CONSOLE="/dev/console"
 mkdir -p /mnt
 dual_boot=$([ -f /sys/firmware/devicetree/base/mediatek,dual-boot ] && echo Y)
+no_split_fitrw=$([ -f /sys/firmware/devicetree/base/mediatek,no-split-fitrw ] && echo Y)
 
-if [ x"${dual_boot}" = xY ]; then
+if [ x"${dual_boot}" = xY -a x"${no_split_fitrw}" = xY ]; then
 	data_dev=$(blkid -t "PARTLABEL=rootfs_data" -o device)
 
 	mkfs.f2fs -q -l rootfs_data $data_dev
